requirements.txt
faiss-cpu==1.8.0.post1
fastapi==0.114.1
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
tiktoken==0.7.0
numpy==1.26.4
orjson==3.10.7
openai==1.45.0